Add files via upload
This is the work flow of the system I've created. This is meant to be kept as general as possible due to the vast possibilities that it enables. Your data may simply be pricing data or it could be quarterly earnings reports, amongst a vast universe of other things. Your buying/selling signals will then change as well. Regardless of what your strategy is, the experimental framework allows you to plug in your data, your signal, and assess the results.